add different background color modes, and an "extra patch"
Idea is to diagnose problems in the stereoscopic setup. The modes are:
0: BG_NONE: none/disabled
1: BG_CONSTANT: constant color (gray)
2: BG_PER_FRAME: cycle through colors per frame
3: BG_STEREO_PER_EYE: use a different color per eye

Each mode can be selected individuallz for the main background, as well
as an "extra Patch". This is meant to be used as follows:
   --stereo --backgroundMode 3 --extraPatch 2
to get a stereo setup with clear left/right eye encoding, and to use
the small patch as synchronization test.